pyproject.toml
[project]
name = "agentic-ai-day"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen>=0.9.7",
    "faiss-cpu",
    "google-adk>=1.8.0",
    "pymupdf>=1.26.3",
    "pypdf2",
    "sentence-transformers",
]
